+++ app/src/main/java/com/example/xmlassignment/Student.kt
@@ -1,3 +1,3 @@
 package com.example.xmlassignment
 
-class Student (val id :Int , val name :String , val grade :Int )
+class Student (val id :Int , val name :String , val grade :Double )
